chore: movido para a pasta UserAccountControllers para separar e mudança no namespace
